Update/57 dto 및 변수 변경 및 드래그앤 드롭 오류 해결

+++ JeJu/src/main/java/com/kgitbank/jeju/dto/FamousRestaurant.java
@@ -27,8 +27,8 @@ public class FamousRestaurant {
 	private Date registration_time;
 	
 	private String location;
-	private Float latitude;
-	private Float longitude;
+	private double latitude;
+	private double longitude;
